Change encoding enforced by Jabref.
The encoding is changed from ISO-8859-1 (
https://en.wikipedia.org/wiki/ISO/IEC_8859-1 ), which supports some
special characters, to ASCII ( https://en.wikipedia.org/wiki/ASCII ),
which does not. The intent is that Jabref will not allow files with
special characters to be saved if the encoding does not allow it. This
should help with detection if some anomalous characters slip into an
entry (e.g. if entries are downloaded from DOI, e-print, etc.).